FIXED IT BY USING 32 BIT NOT 16 IDK WHERE I WAS OVERFLOWING

+++ shorttype.h
@@ -35,27 +35,27 @@ by Hasan Sabri, DO NOT REMOVE CREDIT
 #define sta static
 
 typedef struct {
-	double x, y;
+	f32 x, y;
 } v2;
 typedef struct {
-	int x, y;
+	i32 x, y;
 } v2i;
 
 typedef struct {
-	double x, y, z;
+	f32 x, y, z;
 } v3;
 typedef struct {
-	int x, y, z;
+	i32 x, y, z;
 } v3i;
 
 
 #ifdef SHORTTYPE
-int clampi(int in, int min, int max) {
+i32 clampi(i32 in, i32 min, i32 max) {
 	if (in > max) return max;
 	if (in < min) return min;
 	return in;
 }
-double clampf(double in, double min, double max) {
+f32 clampf(f32 in, f32 min, f32 max) {
 	if (in > max) return max;
 	if (in < min) return min;
 	return in;
@@ -63,7 +63,7 @@ double clampf(double in, double min, double max) {
 
 
 v2 v2_norm(v2 v) {
-	double h = (v.x > v.y) ? v.x : v.y;
+	f32 h = (v.x > v.y) ? v.x : v.y;
 	if (!h) h = 0.00000001;
 	return (v2) {
 		v.x/h,
@@ -71,34 +71,34 @@ v2 v2_norm(v2 v) {
 	};
 }
 
-v2 v2_rot(v2 v, double a) {
-	double sa = sin(a);
-	double ca = cos(a);
+v2 v2_rot(v2 v, f32 a) {
+	f32 sa = sin(a);
+	f32 ca = cos(a);
 	return (v2) {
 		 v.x * ca - v.y * sa,
 		-v.x * sa - v.y * ca
 	};
 }
-v2i v2i_rot(v2i v, double a) {
-	double sa = sin(a);
-	double ca = cos(a);
+v2i v2i_rot(v2i v, f32 a) {
+	f32 sa = sin(a);
+	f32 ca = cos(a);
 	return (v2i) {
 		 v.x * ca - v.y * sa,
 		-v.x * sa - v.y * ca
 	};
 }
-double v2_dis(v2 v) {
+f32 v2_dis(v2 v) {
 	return sqrt(
 		(v.x * v.x) + (v.y * v.y)
 	);
 }
 
-double lerpf(double in, double target, double factor) {
+f32 lerpf(f32 in, f32 target, f32 factor) {
 	return in * (1.0f - factor) + (target * factor);
 }
 
 // v3 v3_norm(v3 v) {
-	// double h = (v.x > v.y) ? v.x : v.y;
+	// f32 h = (v.x > v.y) ? v.x : v.y;
 	// if (h < v.z) h = v.z;
 	// return (v3) {
 		// v.x/h,
@@ -106,9 +106,9 @@ double lerpf(double in, double target, double factor) {
 		// v.z/h
 	// };
 // }
-/*v3 v3_rotx(v3 v, double a) {
-	double sa = sin(a);
-	double ca = cos(a);
+/*v3 v3_rotx(v3 v, f32 a) {
+	f32 sa = sin(a);
+	f32 ca = cos(a);
 	
 	return (v3) {
 		v.x * ca - v.y * sa,
@@ -116,9 +116,9 @@ double lerpf(double in, double target, double factor) {
 		0
 	};
 }
-v3 v3_roty(v3 v, double a) {
-	double sa = sin(a);
-	double ca = cos(a);
+v3 v3_roty(v3 v, f32 a) {
+	f32 sa = sin(a);
+	f32 ca = cos(a);
 	
 	return (v3) {
 		v.x * ca - v.y * sa,
@@ -126,9 +126,9 @@ v3 v3_roty(v3 v, double a) {
 		0
 	};
 }*/
-v3 v3_rotz(v3 v, double a) {
-	double sa = sin(a);
-	double ca = cos(a);
+v3 v3_rotz(v3 v, f32 a) {
+	f32 sa = sin(a);
+	f32 ca = cos(a);
 	
 	return (v3) {
 		v.x * ca - v.y * sa,
@@ -137,18 +137,18 @@ v3 v3_rotz(v3 v, double a) {
 	};
 }
 #else
-int clampi(int in, int min, int max);
-double clampf(double in, double min, double max);
-double lerpf(double in, double target, double factor);
+i32 clampi(i32 in, i32 min, i32 max);
+f32 clampf(f32 in, f32 min, f32 max);
+f32 lerpf(f32 in, f32 target, f32 factor);
 
 v2 v2_norm(v2 v);
-v2 v2_rot(v2 v, double a);
-double v2_dis(v2 v);
+v2 v2_rot(v2 v, f32 a);
+f32 v2_dis(v2 v);
 v3 v3_norm(v3 v);
-v3 v3_rotx(v3 v, double a);
-v3 v3_roty(v3 v, double a);
-v3 v3_rotz(v3 v, double a);
-v2i v2i_rot(v2i v, double a);
+v3 v3_rotx(v3 v, f32 a);
+v3 v3_roty(v3 v, f32 a);
+v3 v3_rotz(v3 v, f32 a);
+v2i v2i_rot(v2i v, f32 a);
 
 
 #endif
